macOS: Don't override action set during drop handling
When the drop has been handled, the target might have accepted a specific
action by calling QDropEvent::setDropAction. Don't override that with
the operation received by the OS if the drag'n'drop operation takes place
within the same application. If the operation comes from outside, we have
no choice but to trust the OS.

This way the drag-site will get the action accepted by the drop-site when
QDrag::drag returns.

Fixes: QTBUG-77427
Pick-to: 6.1 6.0 5.15
Change-Id: I0e4c86adeeea66b345966a03a3251fa62263b3e8
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>