Update commercial license headers to \bin, \mkspecs, \qmake
Update header.COMM to the files in tqtc-qtbase\bin,
tqtc-qtbase\mkspecs and tqtc-qtbase\qmake.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the line count causes failures in tests.

Task-number: QTQAINFRA-4154
Change-Id: I00a45e3989416057c28c90d1107c27ee0b8cbf29
Reviewed-by: Jani Heikkinen <jani.heikkinen@qt.io>